refactor: DeliveryInfoRequest 객체를 공통적으로 사용 못하는 상황이 나와 삭제(Validation 때문)
